+++ sql/core/src/test/scala/org/apache/spark/sql/DataFrameJoinSuite.scala
@@ -98,36 +98,6 @@ class DataFrameJoinSuite extends QueryTest with SharedSQLContext {
       Row(3, "3", 4) :: Nil)
   }
 
-  test("join - join using self join") {
-    val df = Seq(1, 2, 3).map(i => (i, i.toString)).toDF("int", "str")
-
-    // self join
-    checkAnswer(
-      df.join(df, "int"),
-      Row(1, "1", "1") :: Row(2, "2", "2") :: Row(3, "3", "3") :: Nil)
-  }
-
-  test("join - self join") {
-    val df1 = testData.select(testData("key")).as('df1)
-    val df2 = testData.select(testData("key")).as('df2)
-
-    checkAnswer(
-      df1.join(df2, $"df1.key" === $"df2.key"),
-      sql("SELECT a.key, b.key FROM testData a JOIN testData b ON a.key = b.key")
-        .collect().toSeq)
-  }
-
-  test("join - self join auto resolve ambiguity with case insensitivity") {
-    val df = Seq((1, "1"), (2, "2")).toDF("key", "value")
-    checkAnswer(
-      df.join(df, df("key") === df("Key")),
-      Row(1, "1", 1, "1") :: Row(2, "2", 2, "2") :: Nil)
-
-    checkAnswer(
-      df.join(df.filter($"value" === "2"), df("key") === df("Key")),
-      Row(2, "2", 2, "2") :: Nil)
-  }
-
   test("join - cross join") {
     val df1 = Seq((1, "1"), (3, "3")).toDF("int", "str")
     val df2 = Seq((2, "2"), (4, "4")).toDF("int", "str")
@@ -141,76 +111,6 @@ class DataFrameJoinSuite extends QueryTest with SharedSQLContext {
       df2.crossJoin(df1),
       Row(2, "2", 1, "1") :: Row(2, "2", 3, "3") ::
         Row(4, "4", 1, "1") :: Row(4, "4", 3, "3") :: Nil)
-  }
-
-  test("join - using aliases after self join") {
-    val df = Seq(1, 2, 3).map(i => (i, i.toString)).toDF("int", "str")
-    checkAnswer(
-      df.as('x).join(df.as('y), $"x.str" === $"y.str").groupBy("x.str").count(),
-      Row("1", 1) :: Row("2", 1) :: Row("3", 1) :: Nil)
-
-    checkAnswer(
-      df.as('x).join(df.as('y), $"x.str" === $"y.str").groupBy("y.str").count(),
-      Row("1", 1) :: Row("2", 1) :: Row("3", 1) :: Nil)
-  }
-
-  test("[SPARK-6231] join - self join auto resolve ambiguity") {
-    val df = Seq((1, "1"), (2, "2")).toDF("key", "value")
-    checkAnswer(
-      df.join(df, df("key") === df("key")),
-      Row(1, "1", 1, "1") :: Row(2, "2", 2, "2") :: Nil)
-
-    checkAnswer(
-      df.join(df.filter($"value" === "2"), df("key") === df("key")),
-      Row(2, "2", 2, "2") :: Nil)
-
-    checkAnswer(
-      df.join(df, df("key") === df("key") && df("value") === 1),
-      Row(1, "1", 1, "1") :: Nil)
-
-    val left = df.groupBy("key").agg(count("*"))
-    val right = df.groupBy("key").agg(sum("key"))
-    checkAnswer(
-      left.join(right, left("key") === right("key")),
-      Row(1, 1, 1, 1) :: Row(2, 1, 2, 2) :: Nil)
-  }
-
-  test("SPARK-28344: fail ambiguous self join") {
-    val df1 = spark.range(3)
-    val df2 = df1.filter($"id" > 0)
-
-    withSQLConf(
-      SQLConf.FAIL_AMBIGUOUS_SELF_JOIN.key -> "false",
-      SQLConf.CROSS_JOINS_ENABLED.key -> "true") {
-      // `df1("id") > df2("id")` is always false.
-      checkAnswer(df1.join(df2, df1("id") > df2("id")), Nil)
-
-      // `df2("id")` actually points to the column of `df1`.
-      checkAnswer(df1.join(df2).select(df2("id")), Seq(0, 0, 1, 1, 2, 2).map(Row(_)))
-
-      val df3 = df1.filter($"id" <= 2)
-      // `df2("id") < df3("id")` is always false
-      checkAnswer(df1.join(df2).join(df3, df2("id") < df3("id")), Nil)
-    }
-
-    withSQLConf(
-      SQLConf.FAIL_AMBIGUOUS_SELF_JOIN.key -> "true",
-      SQLConf.CROSS_JOINS_ENABLED.key -> "true") {
-
-      def assertAmbiguousSelfJoin(df: => DataFrame): Unit = {
-        val e = intercept[AnalysisException](df)
-        assert(e.message.contains("ambiguous"))
-      }
-
-      assertAmbiguousSelfJoin(df1.join(df2, df1("id") > df2("id")))
-      assertAmbiguousSelfJoin(df1.join(df2).select(df2("id")))
-      assertAmbiguousSelfJoin(df1.join(df2).join(df1, df1("id") > df2("id")))
-      assertAmbiguousSelfJoin(df1.join(df2).join(df1, df2("id") === 2))
-
-      val df3 = df1.filter($"id" <= 2)
-      assertAmbiguousSelfJoin(df1.join(df2).join(df3, df2("id") < df3("id")))
-      assertAmbiguousSelfJoin(df3.join(df1.join(df2), df2("id") < df3("id")))
-    }
   }
 
   test("broadcast join hint using broadcast function") {
